Move InputValue, GraphQLError, and CacheReference back to Apollo
TODO: Fix unneeded import statements and `JSONObject` -> `[String: Any]` conversion
