+++ Sources/ApolloModels/CodegenV1/ScalarTypes.swift
@@ -2,5 +2,6 @@ public protocol ScalarType {}
 
 extension String: ScalarType {}
 extension Int: ScalarType {}
-extension Bool: ScalarType {}
 extension Float: ScalarType {}
+extension Double: ScalarType {}
+extension Bool: ScalarType {}
