+++ Sources/ApolloModels/InputValue.swift
@@ -6,7 +6,7 @@ import Foundation
 public indirect enum InputValue {
   /// A direct input value, valid types are `String`, `Int` `Float` and `Bool`.
   /// For enum input values, the enum cases's `rawValue` as a `String` should be used.
-  case scalar(JSONEncodable)
+  case scalar(ScalarType)
 
   /// A variable input value to be evaluated using the operation's `variables` dictionary at runtime.
   case variable(String)
@@ -22,28 +22,6 @@ public indirect enum InputValue {
   /// A null input value.
   /// - See: [GraphQLSpec - Input Values - Null Value](http://spec.graphql.org/June2018/#sec-Null-Value)
   case none
-
-  public func evaluate(with variables: [String: JSONEncodable]?) throws -> JSONValue {
-    switch self {
-    case .scalar(let scalar):
-      return scalar.jsonValue
-
-    case .variable(let name):
-      guard let value = variables?[name] else {
-        throw GraphQLError("Variable \(name) was not provided.")
-      }
-      return value.jsonValue
-
-    case .list(let array):
-      return try evaluate(values: array, with: variables)
-
-    case .object(let dictionary):
-      return try evaluate(values: dictionary, with: variables)
-
-    case .none:
-      return NSNull()
-    }
-  }
 }
 
 extension InputValue: ExpressibleByNilLiteral {
@@ -80,33 +58,10 @@ extension InputValue: ExpressibleByArrayLiteral {
   @inlinable public init(arrayLiteral elements: InputValue...) {
     self = .list(Array(elements))
   }
-
-  private func evaluate(values: [InputValue], with variables: [String: JSONEncodable]?) throws -> JSONValue {
-    return try evaluate(values: values, with: variables) as [JSONValue]
-  }
-
-  private func evaluate(values: [InputValue], with variables: [String: JSONEncodable]?) throws -> [JSONValue] {
-    try values.map { try $0.evaluate(with: variables) }
-  }
 }
 
 extension InputValue: ExpressibleByDictionaryLiteral {
   public init(dictionaryLiteral elements: (String, InputValue)...) {
     self = .object(Dictionary(elements, uniquingKeysWith: { $1 }))
-  }
-
-  private func evaluate(values: [String: InputValue], with variables: [String: JSONEncodable]?) throws -> JSONValue {
-    return try evaluate(values: values, with: variables) as JSONObject
-  }
-
-  private func evaluate(values: [String: InputValue], with variables: [String: JSONEncodable]?) throws -> JSONObject {
-    var jsonObject = JSONObject(minimumCapacity: values.count)
-    for (key, value) in values {
-      let evaluatedValue = try value.evaluate(with: variables)
-      if !(evaluatedValue is NSNull) {
-        jsonObject[key] = evaluatedValue
-      }
-    }
-    return jsonObject
   }
 }
